+++ riot/lor/model.go
@@ -3,6 +3,6 @@ package lor
 // Player represents a ranked Legends of Runeterra player.
 type Player struct {
 	Name         string `json:"name"`
-	Rank         int    `json:"rank"`
-	LeaguePoints int    `json:"lp"`
+	Rank         int64  `json:"rank"`
+	LeaguePoints int64  `json:"lp"`
 }
